+++ CHANGELOG.md
@@ -14,6 +14,7 @@ straightforward as possible.
 
 - async load of camera sensor images for admin [#532]
 - endpoints for network sync [#548]
+- network synchronization [#549]
 
 ### Changed